% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{bin}
\alias{bin}
\title{Bin values into increasingly wider bins}
\usage{
bin(x)
}
\arguments{
\item{x}{values}
}
\value{
rebinned values
}
\description{
Bin values into increasingly wider bins
}
