% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{get_average_precision_p_value}
\alias{get_average_precision_p_value}
\title{Get average precision p-value}
\usage{
get_average_precision_p_value(nulls, m, n, statistic)
}
\arguments{
\item{nulls}{Nulls data frame}

\item{m}{Number of positive examples (= number of replicates - 1)}

\item{n}{Number of negative examples (= number of controls, or number of non-replicates)}

\item{statistic}{Average precision statistic}
}
\value{
p-value
}
\description{
Get average precision p-value
}
