% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{get_p_value}
\alias{get_p_value}
\title{Compute p-value against null distribution}
\usage{
get_p_value(nulls, m, n, statistic, metric_name = "average_precision")
}
\arguments{
\item{nulls}{Null distribution data frame}

\item{m}{Number of positive examples}

\item{n}{Number of negative examples}

\item{statistic}{statistic for which to compute p-value}

\item{metric_name}{name of metric. Only \code{"average_precision"} is implemented.}
}
\value{
p-value
}
\description{
Compute p-value against null distribution
}
