% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matric.R
\docType{package}
\name{matric}
\alias{matric}
\title{matric: Metrics for Similarity Matrices}
\description{
This is a prototype for testing out ideas for
\code{cytominer-eval} \url{https://github.com/cytomining/cytominer-eval}.
}
