% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_validate_sim.R
\name{new_sim}
\alias{new_sim}
\title{Constructor for \code{sim} S3 class.}
\usage{
new_sim(x, row_metadata, metric_metadata)
}
\arguments{
\item{x}{tbl with similarity matrix.}

\item{row_metadata}{tbl with row metadata.}

\item{metric_metadata}{list with metric information}
}
\value{
object of class \code{sim}
}
\description{
\code{new_sim} creates an object of class \code{sim}.
}
