% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{null_distribution_helper}
\alias{null_distribution_helper}
\title{Compute null distribution of metrics}
\usage{
null_distribution_helper(m, n, nn = 10000, metric_name = "average_precision")
}
\arguments{
\item{m}{Number of positive examples (= number of replicates - 1)}

\item{n}{Number of negative examples (= number of controls, or number of non-replicates)}

\item{nn}{Number of simulations (default = 10000)}

\item{metric_name}{name of metric. Only \code{"average_precision"} is implemented.}
}
\value{
Null distribution data frame parametrized by m and n
}
\description{
Compute null distribution of metrics
}
