% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{retrieval_baseline}
\alias{retrieval_baseline}
\title{Compute null thresholds for a set of metrics}
\usage{
retrieval_baseline(
  metrics,
  background_type,
  level_identifier,
  n_iterations = 10000,
  random_seed = 42
)
}
\arguments{
\item{metrics}{Metrics data frame, containing columns
\verb{sim_stat_signal_n_\{background_type\}_\{level_identifier\}} and
\verb{sim_stat_background_n_\{background_type\}_\{level_identifier\}}}

\item{background_type}{Background type. Either "ref" or "non_rep"}

\item{level_identifier}{Level identifier. Either "i" (Level 1_0) or "g" (Level 2_1)}

\item{random_seed}{Random seed (default = 42)}
}
\value{
Nulls data frame
}
\description{
Compute null thresholds for a set of metrics
}
