% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{retrieval_baseline_helper}
\alias{retrieval_baseline_helper}
\title{Estimate statistics of the distribution of information retrieval metrics under the null hypothesis}
\usage{
retrieval_baseline_helper(m, n, nn = 10000)
}
\arguments{
\item{m}{Number of positive examples (= number of replicates - 1)}

\item{n}{Number of negative examples (= number of controls, or number of non-replicates)}

\item{nn}{Number of simulations (default = 10000)}
}
\value{
Nulls data frame parametrized by m and n
}
\description{
Estimate statistics of the distribution of information retrieval metrics under the null hypothesis
}
