% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_collate.R
\name{sim_collate}
\alias{sim_collate}
\title{Filter rows of the melted similarity matrix to create several sets of pairs.}
\usage{
sim_collate(
  sim_df,
  all_same_cols_rep,
  annotation_cols,
  all_same_cols_ref = NULL,
  all_same_cols_rep_ref = NULL,
  any_different_cols_non_rep = NULL,
  all_same_cols_non_rep = NULL,
  all_different_cols_non_rep = NULL,
  any_different_cols_group = NULL,
  all_same_cols_group = NULL,
  reference = NULL,
  drop_reference = FALSE,
  drop_group = NULL
)
}
\arguments{
\item{sim_df}{tbl with melted similarity matrix.}

\item{all_same_cols_rep}{optional character vector specifying columns.}

\item{annotation_cols}{character vector specifying which columns from \code{metadata} to annotate the left index of the filtered \code{sim_df} with.}

\item{all_same_cols_ref}{optional character vector specifying columns.}

\item{all_same_cols_rep_ref}{optional character vector specifying columns.}

\item{any_different_cols_non_rep}{optional character vector specifying columns.}

\item{all_same_cols_non_rep}{optional character vector specifying columns.}

\item{all_different_cols_non_rep}{optional character vector specifying columns.}

\item{any_different_cols_group}{optional character vector specifying columns.}

\item{all_same_cols_group}{optional character vector specifying columns.}

\item{reference}{optional character string specifying reference.}

\item{drop_reference}{optional boolean specifying whether to filter (drop) pairs using \code{reference} on the left index.}

\item{drop_group}{optional tbl; rows that match on \code{drop_group} on the left or right index are dropped.}
}
\value{
filtered \code{sim_df} with sets of pairs.
}
\description{
\code{sim_some_different_drop_some} Filters melted similarity matrix to create several sets of pairs.
}
\details{
0. Filter out some rows

Filter out pairs that match \code{drop_group} in either right or left indices

1. Similarity to reference

Fetch similarities between
a. all rows (except, optionally those containing \code{reference})
and
b. all rows containing \code{reference}
Do so only for those (a, b) pairs that
- have *same* values in *all* columns of \code{all_same_cols_ref}

2. Similarity to replicates (no references)

Fetch similarities between
a. all rows except \code{reference} rows
and
b. all rows except \code{reference} rows (i.e. to each other)

Do so for only those (a, b) pairs that
- have *same* values in *all* columns of \code{all_same_cols_rep}

Keep, both, (a, b) and (b, a)

3. Similarity to replicates (only references)

Fetch similarities between
a. all rows containing \code{reference}
and
b. all rows containing \code{reference} (i.e. to each other)

Do so for only those (a, b) pairs that
- have *same* values in *all* columns of \code{all_same_cols_rep_ref}.

Keep, both, (a, b) and (b, a)

4. Similarity to non-replicates

Fetch similarities between
a. all rows (except, optionally, \code{reference} rows)
and
b. all rows except \code{reference} rows

Do so for only those (a, b) pairs that
- have *same* values in *all* columns of \code{all_same_cols_non_rep}
- have *different* values in *all* columns \code{all_different_cols_non_rep}
- have *different* values in *at least one* column of \code{any_different_cols_non_rep}

Keep, both, (a, b) and (b, a)

5. Similarity to group

Fetch similarities between
a. all rows (except, optionally, \code{reference} rows)
and
b. all rows (except, optionally, \code{reference} rows)

Do so for only those (a, b) pairs that
- have *same* values in *all* columns of \code{all_same_cols_group}
- have *different* values in *at least one* column of \code{any_different_cols_group}

Keep, both, (a, b) and (b, a)
}
