% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{sim_metrics}
\alias{sim_metrics}
\title{Compute metrics.}
\usage{
sim_metrics(
  munged_sim,
  sim_type,
  calculate_grouped = FALSE,
  annotation_prefix = "Metadata_"
)
}
\arguments{
\item{munged_sim}{output of \code{sim_collated}.}

\item{sim_type}{character string specifying the background distributions for computing scaled metrics. This must be one of the strings \code{"non_rep"} or \code{"ref"}.}

\item{calculate_grouped}{optional boolean specifying whether to include grouped metrics.}

\item{annotation_prefix}{optional character string specifying prefix for annotation columns (e.g. \code{"Metadata_"} (default)).}
}
\value{
list of metrics.
}
\description{
\code{sim_metrics} computes metrics.
}
