% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_construct-validate.R
\name{sim_new}
\alias{sim_new}
\title{Constructor for \code{sim} S3 class.}
\usage{
sim_new(x, row_metadata, metric_metadata)
}
\arguments{
\item{x}{tbl with similarity matrix.}

\item{row_metadata}{tbl with row metadata.}

\item{metric_metadata}{list with metric information}
}
\value{
object of class \code{sim}
}
\description{
\code{sim_new} creates an object of class \code{sim}.
}
