% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_construct-validate.R
\name{sim_preserve}
\alias{sim_preserve}
\title{Preserver for \code{sim} S3 class.}
\usage{
sim_preserve(x, x0)
}
\arguments{
\item{x}{object to preserve.}

\item{x0}{object of class \code{sim}.}
}
\value{
object of class \code{sim} if \code{x} is a valid object of that class
}
\description{
\code{sim_preserve} preserves the attributes of class \code{sim}.
}
\details{
This is a workaround until we figure out S3 inheritance better
https://adv-r.hadley.nz/s3.html#inheritance
}
