% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_construct-validate.R
\name{sim_validate}
\alias{sim_validate}
\title{Validator for \code{matric_sim} S3 class.}
\usage{
sim_validate(x)
}
\arguments{
\item{x}{object.}
}
\value{
Object of class \code{matric_sim} if \code{x} is a valid object of
that class.
}
\description{
\code{sim_validate} validates that an object is of class class \code{matric_sim}.
}
