% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_validate_sim.R
\name{validate_sim}
\alias{validate_sim}
\title{Validator for \code{sim} S3 class.}
\usage{
validate_sim(x)
}
\arguments{
\item{x}{object.}
}
\value{
object of class \code{sim} if \code{x} is a valid object of that class
}
\description{
\code{validate_sim} validates that an object is of class class \code{sim}.
}
