% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellhealthmetrics.R
\docType{data}
\name{cellhealthmetrics}
\alias{cellhealthmetrics}
\title{Quality metrics for Cell Painting dataset of CRISPR perturbations.}
\format{
A list of data frames comprising quality metrics at three levels.
}
\source{
\url{https://github.com/broadinstitute/cell-health/}
}
\usage{
cellhealthmetrics
}
\description{
Quality metrics for  Cell Painting dataset of CRISPR perturbations from
\url{https://www.molbiolcell.org/doi/abs/10.1091/mbc.E20-12-0784}
}
\keyword{datasets}
