% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_calculate.R
\name{sim_calculate}
\alias{sim_calculate}
\title{Calculate a melted similarity matrix.}
\usage{
sim_calculate(
  population,
  annotation_prefix = "Metadata_",
  strata = NULL,
  method = "pearson",
  lazy = FALSE
)
}
\arguments{
\item{population}{data.frame with annotations (a.k.a. metadata) and
observation variables.}

\item{annotation_prefix}{optional character string specifying prefix
for annotation columns.}

\item{strata}{optional character vector specifying stratification columns.}

\item{method}{optional character string specifying method for
to calculate similarity. This must be one of the
strings \code{"pearson"} (default), \code{"kendall"}, \code{"spearman"},
\code{"euclidean"}, \code{"cosine"}.}

\item{lazy}{optional boolean specifying whether to lazily evaluate
similarity.}
}
\value{
\code{metric_sim} object, with similarity matrix and related metadata
}
\description{
\code{sim_calculate} calculates a melted similarity matrix.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tribble(
  ~Metadata_group, ~x, ~y, ~z,
  1, -1, 5, -5,
  2, 0, 6, -4,
  3, 7, -4, 3,
  4, 14, -8, 6
)
sim_pearson <- matric::sim_calculate(population, method = "pearson")
sim_cosine <- matric::sim_calculate(population, method = "cosine")
sim_euclidean <- matric::sim_calculate(population, method = "euclidean")

sim_pearson \%>\%
  dplyr::inner_join(sim_cosine,
    by = c("id1", "id2"),
    suffix = c("_pearson", "_cosine")
  ) \%>\%
  dplyr::inner_join(sim_euclidean \%>\% dplyr::rename(sim_euclidean = sim),
    by = c("id1", "id2")
  )
}
