% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_calculate.R
\name{sim_calculate}
\alias{sim_calculate}
\title{Calculate melted similarity matrix.}
\usage{
sim_calculate(population, annotation_prefix = "Metadata_", method = "pearson")
}
\arguments{
\item{population}{tbl with annotations (a.k.a. metadata) and observation variables.}

\item{annotation_prefix}{optional character string specifying prefix for annotation columns.}

\item{method}{optional character string specifying method for \code{stats::cor} to calculate similarity.  This must be one of the strings \code{"pearson"} (default), \code{"kendall"}, \code{"spearman"}.}
}
\value{
data.frame of melted similarity matrix.
}
\description{
\code{sim_calculate} calculates a melted similarity matrix.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 4, replace = TRUE),
  x = rnorm(4),
  y = x + rnorm(4) / 100,
  z = y + rnorm(4) / 1000
)
matric::sim_calculate(population, method = "pearson")
}
