% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_calculate.R
\name{sim_calculate_ij}
\alias{sim_calculate_ij}
\title{Calculate similarities given pairs of rows}
\usage{
sim_calculate_ij(
  population,
  rows,
  annotation_prefix = "Metadata_",
  method = "cosine",
  cores = 1
)
}
\arguments{
\item{population}{data.frame with annotations (a.k.a. metadata) and
observation variables.}

\item{rows}{data.frame with at least two columns \code{id1} and \code{id2}
specifying rows of \code{population}.}

\item{annotation_prefix}{optional character string specifying prefix
for annotation columns.}

\item{method}{optional character string specifying method for
to calculate similarity. Currently only \code{"cosine"} is implemented.}
}
\value{
\code{rows}, with new column \code{sim} containing similarities.
}
\description{
\code{sim_calculate_ij} calculates similarities given pairs of rows.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tribble(
  ~Metadata_group, ~x, ~y, ~z,
  1, -1, 5, -5,
  2, 0, 6, -4,
  3, 7, -4, 3,
  4, 14, -8, 6
)

n <- nrow(population)

rows <- expand.grid(id1 = seq(n), id2 = seq(n), KEEP.OUT.ATTRS = FALSE)

sim_cosine <- matric::sim_calculate_ij(population, rows, method = "cosine")

sim_cosine
}
