% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics_signif.R
\name{sim_metrics_signif}
\alias{sim_metrics_signif}
\title{Report p-values for metrics}
\usage{
sim_metrics_signif(metrics, background_type, level_identifier, ...)
}
\arguments{
\item{metrics}{Metrics data frame, containing columns
\verb{sim_stat_signal_n_\{background_type\}_\{level_identifier\}} and
\verb{sim_stat_background_n_\{background_type\}_\{level_identifier\}}.}

\item{background_type}{Background type. Either "ref" or "non_rep".}

\item{level_identifier}{Level identifier. Either "i" (Level 1_0) or "g" (Level 2_1).}

\item{...}{arguments passed downstream.}
}
\value{
Metrics data frame, containing additional columns
\verb{sim_retrieval_average_precision_\{background_type\}_\{level_identifier\}_nlog10pvalue}
}
\description{
Report p-values for metrics
}
