% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_construct-validate.R
\name{sim_new}
\alias{sim_new}
\title{Constructor for \code{matric_sim} S3 class.}
\usage{
sim_new(x, row_metadata, metric_metadata)
}
\arguments{
\item{x}{tbl with similarity matrix.}

\item{row_metadata}{tbl with row metadata.}

\item{metric_metadata}{list with metric information}
}
\value{
object of class \code{matric_sim}
}
\description{
\code{sim_new} creates an object of class \code{matric_sim}.
}
\details{
\code{matric_sim} is just a \code{data.frame} with two attributes and at least
three columns.

Columns:
\itemize{
\item \code{id1} and \code{id2}: integers, indicating row ids.
\item \code{sim} similarity: double, indicating similarity between the rows.
}

Attributes:
\itemize{
\item \code{row_metadata}: data.frame of row annotations, with \code{id} column.
\item \code{metric_metadata}: information about the similarity metric.
}
}
