% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_read-write.R
\name{sim_read}
\alias{sim_read}
\title{Read similarity matrix.}
\usage{
sim_read(input, file_format = "parquet")
}
\arguments{
\item{input}{character string specifying the input filename or directory.}
}
\value{
tbl with similarity matrix.
}
\description{
\code{sim_read} reads similarity matrix.
}
\details{
With the CSV format, the \code{row_metadata} and \code{metric_metadata}
attributes are read in and added as attributes to the \code{sim_df} tbl.

This is not required for Parquet because it the attributes are saved.
For Parquet format, \code{sim_read} is identical to
\code{arrow::read_parquet}
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 4, replace = TRUE),
  x = rnorm(4),
  y = x + rnorm(4) / 100,
  z = y + rnorm(4) / 1000
)
sim_df <- sim_calculate(population, method = "pearson")
sim_df \%>\% sim_write("/tmp/test", file_format = "csv")
sim_df_csv <- sim_read("/tmp/test", file_format = "csv")
sim_df \%>\% sim_write("/tmp/test.parquet")
sim_df_parquet1 <- sim_read("/tmp/test.parquet")
sim_df \%>\% arrow::write_parquet("/tmp/test.parquet")
sim_df_parquet2 <- arrow::read_parquet("/tmp/test.parquet")
all(sim_df_parquet1 == sim_df_parquet2)
all(sim_df_parquet1 == sim_df_csv)
}
