% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{sim_write}
\alias{sim_write}
\title{Write similarity matrix.}
\usage{
sim_write(sim_df, output, file_format = "parquet")
}
\arguments{
\item{sim_df}{tbl with melted similarity matrix.}

\item{output}{character string specifying the output directory or filename.}

\item{file_format}{character string specify file format. This must be one of \code{csv} or \code{parquet}(default).}
}
\value{
\code{sim_df}.
}
\description{
\code{sim_write} writes similarity matrix.
}
\details{
The output format can be either CSV or Parquet.

With the CSV format, the \code{row_metadata} and \code{metric_metadata}
attributes are saved as separate files.

This is not required for Parquet because it saves the attributes as well.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 4, replace = TRUE),
  x = rnorm(4),
  y = x + rnorm(4) / 100,
  z = y + rnorm(4) / 1000
)
sim_df <- simplyr::sim_calculate(population, method = "pearson")
sim_df \%>\% simplyr::sim_write("/tmp/test", file_format = "csv")
readr::read_csv("/tmp/test/test.csv")
readr::read_csv("/tmp/test/test_metadata.csv")
jsonlite::read_json("/tmp/test/test_metadata.json")
sim_df \%>\% simplyr::sim_write("/tmp/test.parquet")
sim_df_in <- arrow::read_parquet("/tmp/test.parquet")
attr(sim_df_in, "row_metadata")
attr(sim_df_in, "metric_metadata")
}
